requirements.txt
Flask==2.3.3 
Flask-CORS==4.0.0 
joblib==1.3.2 
numpy==1.24.3 
Pillow==10.0.0 
scikit-learn==1.3.0 
opencv-python==4.8.1.78 
tensorflow==2.14.0 
keras==2.14.0